requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
aiogram>=3.0.0
python-dotenv
alembic
aiogram-calendar